+++ widgets/src/_standalone/ProcessOverview/Pagination.svelte
@@ -31,7 +31,7 @@
 </script>
 
 <div
-	class="flex items-center overflow-y-scroll justify-between border-t border-neutral-400 dark:border-neutral-700 px-4 py-3"
+	class="flex items-center justify-between border-t border-neutral-400 dark:border-neutral-700 px-4 py-3"
 >
 	<div class="min-w-[200px] flex items-center text-sm dark:text-gray-300">
 		<!-- Todo find a way to handle placeholders -->